backend/requirements.txt
fastapi==0.121.0
pydantic==2.11.7
python-multipart==0.0.20
pillow==10.4.0
torch==2.4.1
torchvision==0.19.1
numpy==2.0.2
opencv-python==4.10.0.84
pytorch-grad-cam==1.5.1
reportlab==4.1.5
uvicorn==0.30.0
